feat: expose HostMemoryOrVec via API
Since I needed the same primitive also for RustNN we could expose
it in the trtx API. This helper is commonly when you either
get owned engine bytes from TensorRT or from a cache.
